scriptedforms compiling within jupyterlab (as long as sass pull request is accepted)
