+++ .gitignore
@@ -162,3 +162,5 @@ ENV/
 
 # mypy
 .mypy_cache/
+
+scriptedforms/lab-build
