+++ scriptedforms/package.json
@@ -11,8 +11,8 @@
     "jupyterlab",
     "jupyterlab-extension"
   ],
-  "main": "labextension/lib/index.js",
-  "types": "labextension/lib/index.d.ts",
+  "main": "lab-build/main.js",
+  "types": "lab-build/main.d.ts",
   "description": "Interactive forms.",
   "author": "Simon Biggs",
   "scripts": {
@@ -20,7 +20,8 @@
     "lint": "ng lint",
     "build": "rimraf lib && webpack --config webpack.dev.js && yarn build:labextension",
     "build:prod": "rimraf lib && webpack --config webpack.prod.js && yarn build:labextension",
-    "build:labextension": "rimraf labextension/lib && tsc --project labextension",
+    "build:labextension": "rimraf lab-build && yarn copy:labextension && tsc",
+    "copy:labextension": "copyfiles --up 1 src/**/*.html src/**/*.css src/**/*.scss lab-build/",
     "selenium": "./scripts/selenium.sh",
     "test": "yarn --cwd ./e2e e2e",
     "pip:install": "jlpm pip:dev",
@@ -34,10 +35,12 @@
     "bundle:update": "yarn --cwd ../portable-windows-bundle update",
     "bundle:test": "yarn --cwd ../portable-windows-bundle test",
     "bundle:zip": "yarn --cwd ../portable-windows-bundle zip",
-    "watch": "webpack --config webpack.dev.js --watch"
+    "watch": "yarn watch:main & yarn watch:labextension",
+    "watch:main": "webpack --config webpack.dev.js --watch",
+    "watch:labextension": "yarn copy:labextension && tsc --watch"
   },
   "files": [
-    "build/**/*.{d.ts,eot,gif,html,css,jpg,js,js.map,json,png,svg,woff2,ttf}"
+    "lab-build/**/*.{d.ts,eot,gif,html,css,jpg,js,js.map,json,png,svg,woff2,ttf,scss}"
   ],
   "dependencies": {
     "@angular/animations": "5.2.5",
@@ -62,6 +65,7 @@
     "@jupyterlab/theme-light-extension": "^0.15.4",
     "@phosphor/signaling": "^1.2.2",
     "@phosphor/widgets": "^1.5.0",
+    "core-js": "^2.5.4",
     "hammerjs": "^2.0.8",
     "json-stable-stringify": "^1.0.1",
     "markdown-it": "^8.4.0",
