+++ scriptedforms/src/vendors/theme.scss
@@ -24,7 +24,7 @@
 A custom angular material design theme.
 */
 
-@import "../../node_modules/@angular/material/theming";
+@import "~@angular/material/theming";
 
 // Include the common styles for Angular Material. We include this here so that you only
 // have to load a single css file for Angular Material in your app.
