+++ scriptedforms/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "sourceMap": true,
-    "declaration": false,
+    "declaration": true,
     "noImplicitAny": true,
     "noEmitOnError": true,
     "experimentalDecorators": true,
@@ -10,7 +10,7 @@
     "module": "commonjs",
     "moduleResolution": "node",
     "target": "es5",
-    "outDir": "./build",
+    "outDir": "./lab-build",
     "lib": [
       "es2017",
       "dom"
